requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
